requirements.txt
groq==0.6.0
python-dotenv==1.0.1